+++ test/pgproto/auth_test.py
@@ -1,6 +1,6 @@
 import pytest
 import pg8000.dbapi as pg  # type: ignore
-from conftest import Postgres
+from conftest import Postgres, Cluster
 
 
 def test_auth(postgres: Postgres):
@@ -41,3 +41,38 @@ def test_auth(postgres: Postgres):
         pg.DatabaseError, match=f"authentication failed for user '{sha_user}'"
     ):
         pg.Connection(sha_user, password="aaa", host=postgres.host, port=postgres.port)
+
+
+def test_admin_auth(cluster: Cluster):
+    import os
+
+    os.environ["PICODATA_ADMIN_PASSWORD"] = "#AdminX12345"
+
+    cluster.set_config_file(
+        yaml="""
+    cluster:
+        cluster_id: test
+        tier:
+            default:
+    instance:
+        pg:
+            listen: "127.0.0.1:5432"
+            ssl: False
+    """
+    )
+    i1 = cluster.add_instance(wait_online=False)
+    i1.env.update(os.environ)
+    i1.start()
+    i1.wait_online()
+
+    user = "admin"
+    password = os.getenv("PICODATA_ADMIN_PASSWORD")
+
+    # test authentication with a wrong password
+    with pytest.raises(
+        pg.DatabaseError, match=f"authentication failed for user '{user}'"
+    ):
+        pg.Connection(user, password="wrong password", host="127.0.0.1", port=5432)
+
+    conn = pg.Connection(user=user, password=password, host="127.0.0.1", port=5432)
+    conn.close()